Create position.java
Write the method position() which returns the position of the first appearance of a specified digit in a
positive number n. The position of the digit is counted from the right and starts from 1. If the required
digit is not in the number, the method should return -1. For example, position(12315, 1) returns 2 and
position(12, 3) returns -1