Create extractOddDigits.java
Write a method extractOddDigits() which extracts the odd digits from a positive number n, and
combines the odd digits sequentially into a new number. The new number is returned back to the
calling method. If the input number n does not contain any odd digits, then returns -1. For examples, if
n=1234567, then 1357 is returned; and if n=28, then –1 is returned.